#18 - hungnln implement stat match file

+++ src/main/java/com/hungnln/vleague/DTO/PlayerCreateDTO.java
@@ -28,10 +28,10 @@ public class PlayerCreateDTO {
     @NotNull(message = ValidationMessage.DOB_VALID_MESSAGE)
     @Min(ValidationSize.HEIGHTCM_MIN)
     @Max(ValidationSize.HEIGHTCM_MAX)
-    private float heightCm;
+    private int heightCm;
 
     @Min(ValidationSize.WEIGHTKG_MIN)
     @Max(ValidationSize.WEIGHTKG_MAX)
     @NotNull(message = ValidationMessage.DOB_VALID_MESSAGE)
-    private float weightKg;
+    private int weightKg;
 }
